Update build and download steps in release workflow
The release workflow has been updated to use `libjson-c-dev` instead of `json-c` and `json-c-dev` in the build tools setup. Additionally, the path for `json-c` includes and library has been adjusted in the `gcc` command. In the download binaries step, the path has been modified and labels for downloading x86_64 and arm64 binaries were added.